requirements.txt
cohere==5.16.3
fitz==0.0.1.dev2
langchain==0.3.27
langchain_astradb==0.6.0
langchain_community==0.3.27
langchain_core==0.3.74
langchain_groq==0.3.7
numpy==2.3.2
pandas==2.3.1
phoenix==0.9.1
python-dotenv==1.1.1
python_docx==1.2.0
Requests==2.32.4
